refactor: extract props to avatar-group-props
